Moves Nikke Exceptions
