+++ src/main/kotlin/com/hyuse/nikkeManager/domain/exceptions/nikke/NikkeAlreadyExistsException.kt
@@ -1,3 +1,3 @@
-package com.hyuse.nikkeManager.exceptions
+package com.hyuse.nikkeManager.domain.exceptions.nikke
 
 class NikkeAlreadyExistsException (name: String): NikkeException("Nikke with name '$name' already exists")
