+++ src/main/kotlin/com/hyuse/nikkeManager/domain/exceptions/nikke/NikkeCollectionEmptyException.kt
@@ -1,0 +1,4 @@
+package com.hyuse.nikkeManager.domain.exceptions.nikke
+
+class NikkeCollectionEmptyException (): NikkeException("Collection Empty") {
+}
