+++ src/main/kotlin/com/hyuse/nikkeManager/domain/exceptions/nikke/NikkeException.kt
@@ -1,3 +1,3 @@
-package com.hyuse.nikkeManager.exceptions
+package com.hyuse.nikkeManager.domain.exceptions.nikke
 
 sealed class NikkeException(message: String): RuntimeException(message)
