+++ src/main/kotlin/com/hyuse/nikkeManager/domain/exceptions/nikke/NikkeIdNotFoundException.kt
@@ -1,3 +1,3 @@
-package com.hyuse.nikkeManager.exceptions
+package com.hyuse.nikkeManager.domain.exceptions.nikke
 
 class NikkeIdNotFoundException(id: String) : NikkeException("Nikke with id '$id' not found")
