+++ src/main/kotlin/com/hyuse/nikkeManager/domain/exceptions/nikke/NikkeNotFoundException.kt
@@ -1,3 +1,3 @@
-package com.hyuse.nikkeManager.exceptions
+package com.hyuse.nikkeManager.domain.exceptions.nikke
 
 class NikkeNotFoundException(name: String): NikkeException("Nikke with name '$name' not found")
